Move nist and version parsers to separate files. Refactor.
